# docs/swagger.yaml
basePath: /api/v2
definitions:
...
  autocomplete.ACEntry:
    properties:
      id:
        type: integer
      type:
        type: string
      value:
        type: string
    type: object
...
  services.BaseErrorResponse:
    properties:
      error:
        $ref: '#/definitions/services.RespError'
      status:
        type: integer
      updateToken:
        type: boolean
    type: object
  services.RespError:
    properties:
      errorCode:
        type: integer
      errors:
        items:
          type: string
        type: array
      message:
        type: string
    type: object
...
host: wso.williams.edu
info:
  contact:
    email: wso-dev@wso.williams.edu
    name: WSO Dev
  description: API for WSO services like factrak, facebook, dormtrak, course scheduler,
    and others.
  title: WSO API
  version: 2.0.0
paths:
...
  /autocomplete/area-of-study:
    get:
      consumes:
      - application/json
      description: Given an input q, this autocompletes the area of study. Results
        are sorted by levenshtein distance.
      operationId: autocompleteAreaOfStudy
      parameters:
      - description: String to Complete
        in: query
        name: q
        required: true
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/autocomplete.ACEntry'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/services.BaseErrorResponse'
      security:
      - Bearer: []
      summary: Autocomplete area of study
      tags:
      - autocomplete
  /autocomplete/course:
    get:
      consumes:
      - application/json
      description: Given an input q, this autocompletes the course. Results are sorted
        by levenshtein distance.
      operationId: autocompleteCourse
      parameters:
      - description: String to Complete
        in: query
        name: q
        required: true
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/autocomplete.ACEntry'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/services.BaseErrorResponse'
      security:
      - Bearer: []
      summary: Autocomplete course
      tags:
      - autocomplete
  /autocomplete/factrak:
    get:
      consumes:
      - application/json
      description: Given an input q, this autocompletes the professor or course. Results
        are sorted by levenshtein distance.
      operationId: autocompleteFactrak
      parameters:
      - description: String to Complete
        in: query
        name: q
        required: true
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/autocomplete.ACEntry'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/services.BaseErrorResponse'
      security:
      - Bearer: []
      summary: Autocomplete factrak professors and courses
      tags:
      - autocomplete
  /autocomplete/professor:
    get:
      consumes:
      - application/json
      description: Given an input q, this autocompletes the professor. Results are
        sorted by levenshtein distance.
      operationId: autocompleteProfessor
      parameters:
      - description: String to Complete
        in: query
        name: q
        required: true
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/autocomplete.ACEntry'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/services.BaseErrorResponse'
      security:
      - Bearer: []
      summary: Autocomplete professor
      tags:
      - autocomplete
  /autocomplete/tag:
    get:
      consumes:
      - application/json
      description: Given an input q, this autocompletes the tag. Results are sorted
        by levenshtein distance.
      operationId: autocompleteTag
      parameters:
      - description: String to Complete
        in: query
        name: q
        required: true
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/autocomplete.ACEntry'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/services.BaseErrorResponse'
      security:
      - Bearer: []
      summary: Autocomplete tag
      tags:
      - autocomplete
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"